frrender: T6746: move get_frrender_dict from vyos.configdict to this module
Keep all FRRender stuff in one place.
