+++ src/conf_mode/protocols_mpls.py
@@ -20,9 +20,9 @@ from sys import exit
 
 from glob import glob
 from vyos.config import Config
-from vyos.configdict import get_frrender_dict
 from vyos.configverify import has_frr_protocol_in_dict
 from vyos.frrender import FRRender
+from vyos.frrender import get_frrender_dict
 from vyos.utils.dict import dict_search
 from vyos.utils.file import read_file
 from vyos.utils.process import is_systemd_service_running
