+++ src/conf_mode/protocols_pim.py
@@ -22,10 +22,10 @@ from signal import SIGTERM
 from sys import exit
 
 from vyos.config import Config
-from vyos.configdict import get_frrender_dict
 from vyos.configverify import verify_interface_exists
 from vyos.configverify import has_frr_protocol_in_dict
 from vyos.frrender import FRRender
+from vyos.frrender import get_frrender_dict
 from vyos.frrender import pim_daemon
 from vyos.utils.process import is_systemd_service_running
 from vyos.utils.process import process_named_running
