+++ src/conf_mode/protocols_static.py
@@ -19,11 +19,11 @@ from sys import exit
 from sys import argv
 
 from vyos.config import Config
-from vyos.configdict import get_frrender_dict
 from vyos.configverify import has_frr_protocol_in_dict
 from vyos.configverify import verify_common_route_maps
 from vyos.configverify import verify_vrf
 from vyos.frrender import FRRender
+from vyos.frrender import get_frrender_dict
 from vyos.utils.process import is_systemd_service_running
 from vyos.template import render
 from vyos import ConfigError
